re-unify imagedumper_js and imagedumper_notmobile
