--- imagedumper_notmobile.go
@@ -1,150 +1,0 @@
-// Copyright 2018 The Ebiten Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//go:build !android && !ios
-// +build !android,!ios
-
-package ebiten
-
-import (
-	"fmt"
-	"os"
-
-	"github.com/hajimehoshi/ebiten/v2/internal/debug"
-)
-
-func takeScreenshot(screen *Image) error {
-	newname, err := availableFilename("screenshot_", ".png")
-	if err != nil {
-		return err
-	}
-
-	blackbg := !IsScreenTransparent()
-	if err := screen.image.DumpScreenshot(newname, blackbg); err != nil {
-		return err
-	}
-
-	if _, err := fmt.Fprintf(os.Stderr, "Saved screenshot: %s\n", newname); err != nil {
-		return err
-	}
-	return nil
-}
-
-type imageDumper struct {
-	g Game
-
-	keyState map[Key]int
-
-	hasScreenshotKey bool
-	screenshotKey    Key
-	toTakeScreenshot bool
-
-	hasDumpInternalImagesKey bool
-	dumpInternalImagesKey    Key
-	toDumpInternalImages     bool
-
-	err error
-}
-
-func envScreenshotKey() string {
-	if env := os.Getenv("EBITENGINE_SCREENSHOT_KEY"); env != "" {
-		return env
-	}
-	// For backward compatibility, read the EBITEN_ version.
-	return os.Getenv("EBITEN_SCREENSHOT_KEY")
-}
-
-func envInternalImagesKey() string {
-	if env := os.Getenv("EBITENGINE_INTERNAL_IMAGES_KEY"); env != "" {
-		return env
-	}
-	// For backward compatibility, read the EBITEN_ version.
-	return os.Getenv("EBITEN_INTERNAL_IMAGES_KEY")
-}
-
-func (i *imageDumper) update() error {
-	if i.err != nil {
-		return i.err
-	}
-
-	if err := i.g.Update(); err != nil {
-		return err
-	}
-
-	// If keyState is nil, all values are not initialized.
-	if i.keyState == nil {
-		i.keyState = map[Key]int{}
-
-		if keyname := envScreenshotKey(); keyname != "" {
-			if key, ok := keyNameToKeyCode(keyname); ok {
-				i.hasScreenshotKey = true
-				i.screenshotKey = key
-			}
-		}
-
-		if keyname := envInternalImagesKey(); keyname != "" {
-			if debug.IsDebug {
-				if key, ok := keyNameToKeyCode(keyname); ok {
-					i.hasDumpInternalImagesKey = true
-					i.dumpInternalImagesKey = key
-				}
-			} else {
-				fmt.Fprintf(os.Stderr, "EBITENGINE_INTERNAL_IMAGES_KEY is disabled. Specify a build tag 'ebitenginedebug' to enable it.\n")
-			}
-		}
-	}
-
-	keys := map[Key]struct{}{}
-	if i.hasScreenshotKey {
-		keys[i.screenshotKey] = struct{}{}
-	}
-	if i.hasDumpInternalImagesKey {
-		keys[i.dumpInternalImagesKey] = struct{}{}
-	}
-
-	for key := range keys {
-		if IsKeyPressed(key) {
-			i.keyState[key]++
-			if i.keyState[key] == 1 {
-				if i.hasScreenshotKey && key == i.screenshotKey {
-					i.toTakeScreenshot = true
-				}
-				if i.hasDumpInternalImagesKey && key == i.dumpInternalImagesKey {
-					i.toDumpInternalImages = true
-				}
-			}
-		} else {
-			i.keyState[key] = 0
-		}
-	}
-	return nil
-}
-
-func (i *imageDumper) dump(screen *Image) error {
-	if i.toTakeScreenshot {
-		i.toTakeScreenshot = false
-		if err := takeScreenshot(screen); err != nil {
-			return err
-		}
-	}
-
-	if i.toDumpInternalImages {
-		i.toDumpInternalImages = false
-		if err := dumpInternalImages(); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
